+++ ltm.go
@@ -6,196 +6,6 @@ import (
 	"strings"
 )
 
-// ServerSSLProfiles
-// Documentation: https://devcentral.f5.com/wiki/iControlREST.APIRef_tm_ltm_profile_server-ssl.ashx
-
-// ServerSSLProfiles contains a list of every server-ssl profile on the BIG-IP system.
-type ServerSSLProfiles struct {
-	ServerSSLProfiles []ServerSSLProfile `json:"items"`
-}
-
-// ServerSSLProfile contains information about each server-ssl profile. You can use all
-// of these fields when modifying a server-ssl profile.
-type ServerSSLProfile struct {
-	Name                         string   `json:"name,omitempty"`
-	Partition                    string   `json:"partition,omitempty"`
-	FullPath                     string   `json:"fullPath,omitempty"`
-	Generation                   int      `json:"generation,omitempty"`
-	AlertTimeout                 string   `json:"alertTimeout,omitempty"`
-	Authenticate                 string   `json:"authenticate,omitempty"`
-	AuthenticateDepth            int      `json:"authenticateDepth,omitempty"`
-	CaFile                       string   `json:"caFile,omitempty"`
-	CacheSize                    int      `json:"cacheSize,omitempty"`
-	CacheTimeout                 int      `json:"cacheTimeout,omitempty"`
-	Cert                         string   `json:"cert,omitempty"`
-	Chain                        string   `json:"chain,omitempty"`
-	Ciphers                      string   `json:"ciphers,omitempty"`
-	DefaultsFrom                 string   `json:"defaultsFrom,omitempty"`
-	ExpireCertResponseControl    string   `json:"expireCertResponseControl,omitempty"`
-	GenericAlert                 string   `json:"genericAlert,omitempty"`
-	HandshakeTimeout             string   `json:"handshakeTimeout,omitempty"`
-	Key                          string   `json:"key,omitempty"`
-	ModSslMethods                string   `json:"modSslMethods,omitempty"`
-	Mode                         string   `json:"mode,omitempty"`
-	TmOptions                    []string `json:"tmOptions,omitempty"`
-	Passphrase                   string   `json:"passphrase,omitempty"`
-	PeerCertMode                 string   `json:"peerCertMode,omitempty"`
-	ProxySsl                     string   `json:"proxySsl,omitempty"`
-	RenegotiatePeriod            string   `json:"renegotiatePeriod,omitempty"`
-	RenegotiateSize              string   `json:"renegotiateSize,omitempty"`
-	Renegotiation                string   `json:"renegotiation,omitempty"`
-	RetainCertificate            string   `json:"retainCertificate,omitempty"`
-	SecureRenegotiation          string   `json:"secureRenegotiation,omitempty"`
-	ServerName                   string   `json:"serverName,omitempty"`
-	SessionMirroring             string   `json:"sessionMirroring,omitempty"`
-	SessionTicket                string   `json:"sessionTicket,omitempty"`
-	SniDefault                   string   `json:"sniDefault,omitempty"`
-	SniRequire                   string   `json:"sniRequire,omitempty"`
-	SslForwardProxy              string   `json:"sslForwardProxy,omitempty"`
-	SslForwardProxyBypass        string   `json:"sslForwardProxyBypass,omitempty"`
-	SslSignHash                  string   `json:"sslSignHash,omitempty"`
-	StrictResume                 string   `json:"strictResume,omitempty"`
-	UncleanShutdown              string   `json:"uncleanShutdown,omitempty"`
-	UntrustedCertResponseControl string   `json:"untrustedCertResponseControl,omitempty"`
-}
-
-// ClientSSLProfiles
-// Documentation: https://devcentral.f5.com/wiki/iControlREST.APIRef_tm_ltm_profile_client-ssl.ashx
-
-// ClientSSLProfiles contains a list of every client-ssl profile on the BIG-IP system.
-type ClientSSLProfiles struct {
-	ClientSSLProfiles []ClientSSLProfile `json:"items"`
-}
-
-// ClientSSLProfile contains information about each client-ssl profile. You can use all
-// of these fields when modifying a client-ssl profile.
-type ClientSSLProfile struct {
-	Name              string `json:"name,omitempty"`
-	Partition         string `json:"partition,omitempty"`
-	FullPath          string `json:"fullPath,omitempty"`
-	Generation        int    `json:"generation,omitempty"`
-	AlertTimeout      string `json:"alertTimeout,omitempty"`
-	AllowNonSsl       string `json:"allowNonSsl,omitempty"`
-	Authenticate      string `json:"authenticate,omitempty"`
-	AuthenticateDepth int    `json:"authenticateDepth,omitempty"`
-	CaFile            string `json:"caFile,omitempty"`
-	CacheSize         int    `json:"cacheSize,omitempty"`
-	CacheTimeout      int    `json:"cacheTimeout,omitempty"`
-	Cert              string `json:"cert,omitempty"`
-	CertKeyChain      []struct {
-		Name       string `json:"name,omitempty"`
-		Cert       string `json:"cert,omitempty"`
-		Chain      string `json:"chain,omitempty"`
-		Key        string `json:"key,omitempty"`
-		Passphrase string `json:"passphrase,omitempty"`
-	} `json:"certKeyChain,omitempty"`
-	CertExtensionIncludes           []string `json:"certExtensionIncludes,omitempty"`
-	CertLifespan                    int      `json:"certLifespan,omitempty"`
-	CertLookupByIpaddrPort          string   `json:"certLookupByIpaddrPort,omitempty"`
-	Chain                           string   `json:"chain,omitempty"`
-	Ciphers                         string   `json:"ciphers,omitempty"`
-	ClientCertCa                    string   `json:"clientCertCa,omitempty"`
-	CrlFile                         string   `json:"crlFile,omitempty"`
-	DefaultsFrom                    string   `json:"defaultsFrom,omitempty"`
-	ForwardProxyBypassDefaultAction string   `json:"forwardProxyBypassDefaultAction,omitempty"`
-	GenericAlert                    string   `json:"genericAlert,omitempty"`
-	HandshakeTimeout                string   `json:"handshakeTimeout,omitempty"`
-	InheritCertkeychain             string   `json:"inheritCertkeychain,omitempty"`
-	Key                             string   `json:"key,omitempty"`
-	ModSslMethods                   string   `json:"modSslMethods,omitempty"`
-	Mode                            string   `json:"mode,omitempty"`
-	TmOptions                       []string `json:"tmOptions,omitempty"`
-	Passphrase                      string   `json:"passphrase,omitempty"`
-	PeerCertMode                    string   `json:"peerCertMode,omitempty"`
-	ProxyCaCert                     string   `json:"proxyCaCert,omitempty"`
-	ProxyCaKey                      string   `json:"proxyCaKey,omitempty"`
-	ProxyCaPassphrase               string   `json:"proxyCaPassphrase,omitempty"`
-	ProxySsl                        string   `json:"proxySsl,omitempty"`
-	ProxySslPassthrough             string   `json:"proxySslPassthrough,omitempty"`
-	RenegotiatePeriod               string   `json:"renegotiatePeriod,omitempty"`
-	RenegotiateSize                 string   `json:"renegotiateSize,omitempty"`
-	Renegotiation                   string   `json:"renegotiation,omitempty"`
-	RetainCertificate               string   `json:"retainCertificate,omitempty"`
-	SecureRenegotiation             string   `json:"secureRenegotiation,omitempty"`
-	ServerName                      string   `json:"serverName,omitempty"`
-	SessionMirroring                string   `json:"sessionMirroring,omitempty"`
-	SessionTicket                   string   `json:"sessionTicket,omitempty"`
-	SniDefault                      string   `json:"sniDefault,omitempty"`
-	SniRequire                      string   `json:"sniRequire,omitempty"`
-	SslForwardProxy                 string   `json:"sslForwardProxy,omitempty"`
-	SslForwardProxyBypass           string   `json:"sslForwardProxyBypass,omitempty"`
-	SslSignHash                     string   `json:"sslSignHash,omitempty"`
-	StrictResume                    string   `json:"strictResume,omitempty"`
-	UncleanShutdown                 string   `json:"uncleanShutdown,omitempty"`
-}
-
-// Nodes contains a list of every node on the BIG-IP system.
-type Nodes struct {
-	Nodes []Node `json:"items"`
-}
-
-// Node contains information about each individual node. You can use all
-// of these fields when modifying a node.
-type Node struct {
-	Name            string `json:"name,omitempty"`
-	Partition       string `json:"partition,omitempty"`
-	FullPath        string `json:"fullPath,omitempty"`
-	Generation      int    `json:"generation,omitempty"`
-	Address         string `json:"address,omitempty"`
-	ConnectionLimit int    `json:"connectionLimit,omitempty"`
-	DynamicRatio    int    `json:"dynamicRatio,omitempty"`
-	Logging         string `json:"logging,omitempty"`
-	Monitor         string `json:"monitor,omitempty"`
-	RateLimit       string `json:"rateLimit,omitempty"`
-	Ratio           int    `json:"ratio,omitempty"`
-	Session         string `json:"session,omitempty"`
-	State           string `json:"state,omitempty"`
-}
-
-// DataGroups contains a list of data groups on the BIG-IP system.
-type DataGroups struct {
-	DataGroups []DataGroup `json:"items"`
-}
-
-// DataGroups contains information about each data group.
-type DataGroup struct {
-	Name       string
-	Partition  string
-	FullPath   string
-	Generation int
-	Type       string
-	Records    []DataGroupRecord
-}
-
-type DataGroupRecord struct {
-	Name string `json:"name,omitempty"`
-	Data string `json:"data,omitempty"`
-}
-
-type dataGroupDTO struct {
-	Name       string            `json:"name,omitempty"`
-	Partition  string            `json:"partition,omitempty"`
-	FullPath   string            `json:"fullPath,omitempty"`
-	Generation int               `json:"generation,omitempty"`
-	Type       string            `json:"type,omitempty"`
-	Records    []DataGroupRecord `json:"records,omitempty"`
-}
-
-func (p *DataGroup) MarshalJSON() ([]byte, error) {
-	var dto dataGroupDTO
-	marshal(&dto, p)
-	return json.Marshal(dto)
-}
-
-func (p *DataGroup) UnmarshalJSON(b []byte) error {
-	var dto dataGroupDTO
-	err := json.Unmarshal(b, &dto)
-	if err != nil {
-		return err
-	}
-	return marshal(p, &dto)
-}
-
 // SnatPools contains a list of every snatpool on the BIG-IP system.
 type SnatPools struct {
 	SnatPools []SnatPool `json:"items"`
@@ -214,69 +24,6 @@ type SnatPool struct {
 // Pools contains a list of pools on the BIG-IP system.
 type Pools struct {
 	Pools []Pool `json:"items"`
-}
-
-// Pool contains information about each pool. You can use all of these
-// fields when modifying a pool.
-type Pool struct {
-	Name                   string `json:"name,omitempty"`
-	Partition              string `json:"partition,omitempty"`
-	FullPath               string `json:"fullPath,omitempty"`
-	Generation             int    `json:"generation,omitempty"`
-	AllowNAT               string `json:"allowNat,omitempty"`
-	AllowSNAT              string `json:"allowSnat,omitempty"`
-	IgnorePersistedWeight  string `json:"ignorePersistedWeight,omitempty"`
-	IPTOSToClient          string `json:"ipTosToClient,omitempty"`
-	IPTOSToServer          string `json:"ipTosToServer,omitempty"`
-	LinkQoSToClient        string `json:"linkQosToClient,omitempty"`
-	LinkQoSToServer        string `json:"linkQosToServer,omitempty"`
-	LoadBalancingMode      string `json:"loadBalancingMode,omitempty"`
-	MinActiveMembers       int    `json:"minActiveMembers,omitempty"`
-	MinUpMembers           int    `json:"minUpMembers,omitempty"`
-	MinUpMembersAction     string `json:"minUpMembersAction,omitempty"`
-	MinUpMembersChecking   string `json:"minUpMembersChecking,omitempty"`
-	Monitor                string `json:"monitor,omitempty"`
-	QueueDepthLimit        int    `json:"queueDepthLimit,omitempty"`
-	QueueOnConnectionLimit string `json:"queueOnConnectionLimit,omitempty"`
-	QueueTimeLimit         int    `json:"queueTimeLimit,omitempty"`
-	ReselectTries          int    `json:"reselectTries,omitempty"`
-	ServiceDownAction      string `json:"serviceDownAction,omitempty"`
-	SlowRampTime           int    `json:"slowRampTime,omitempty"`
-}
-
-// Pool Members contains a list of pool members within a pool on the BIG-IP system.
-type PoolMembers struct {
-	PoolMembers []PoolMember `json:"items"`
-}
-
-// poolMember is used only when adding members to a pool.
-type poolMember struct {
-	Name string `json:"name"`
-}
-
-// poolMembers is used only when modifying members on a pool.
-type poolMembers struct {
-	Members []PoolMember `json:"members"`
-}
-
-// Pool Member contains information about each individual member in a pool. You can use all
-// of these fields when modifying a pool member.
-type PoolMember struct {
-	Name            string `json:"name,omitempty"`
-	Partition       string `json:"partition,omitempty"`
-	FullPath        string `json:"fullPath,omitempty"`
-	Generation      int    `json:"generation,omitempty"`
-	Address         string `json:"address,omitempty"`
-	ConnectionLimit int    `json:"connectionLimit,omitempty"`
-	DynamicRatio    int    `json:"dynamicRatio,omitempty"`
-	InheritProfile  string `json:"inheritProfile,omitempty"`
-	Logging         string `json:"logging,omitempty"`
-	Monitor         string `json:"monitor,omitempty"`
-	PriorityGroup   int    `json:"priorityGroup,omitempty"`
-	RateLimit       string `json:"rateLimit,omitempty"`
-	Ratio           int    `json:"ratio,omitempty"`
-	Session         string `json:"session,omitempty"`
-	State           string `json:"state,omitempty"`
 }
 
 // VirtualServers contains a list of all virtual servers on the BIG-IP system.
@@ -792,40 +539,6 @@ type Profile struct {
 	Context   string `json:"context,omitempty"`
 }
 
-type IRules struct {
-	IRules []IRule `json:"items"`
-}
-
-type IRule struct {
-	Name      string `json:"name,omitempty"`
-	Partition string `json:"partition,omitempty"`
-	FullPath  string `json:"fullPath,omitempty"`
-	Rule      string `json:"apiAnonymous,omitempty"`
-}
-
-const (
-	uriLtm            = "ltm"
-	uriNode           = "node"
-	uriPool           = "pool"
-	uriPoolMember     = "members"
-	uriProfile        = "profile"
-	uriServerSSL      = "server-ssl"
-	uriClientSSL      = "client-ssl"
-	uriVirtual        = "virtual"
-	uriVirtualAddress = "virtual-address"
-	uriSnatPool       = "snatpool"
-	uriMonitor        = "monitor"
-	uriIRule          = "rule"
-	uriPolicy         = "policy"
-	uriDatagroup      = "data-group"
-	uriInternal       = "internal"
-	ENABLED           = "enable"
-	DISABLED          = "disable"
-	CONTEXT_SERVER    = "serverside"
-	CONTEXT_CLIENT    = "clientside"
-	CONTEXT_ALL       = "all"
-)
-
 var cidr = map[string]string{
 	"0":  "0.0.0.0",
 	"1":  "128.0.0.0",
@@ -912,108 +625,6 @@ func (b *BigIP) DeleteSnatPool(name string) error {
 // can be modified are referenced in the Snatpool struct.
 func (b *BigIP) ModifySnatPool(name string, config *SnatPool) error {
 	return b.put(config, uriLtm, uriSnatPool, name)
-}
-
-// ServerSSLProfiles returns a list of server-ssl profiles.
-func (b *BigIP) ServerSSLProfiles() (*ServerSSLProfiles, error) {
-	var serverSSLProfiles ServerSSLProfiles
-	err, _ := b.getForEntity(&serverSSLProfiles, uriLtm, uriProfile, uriServerSSL)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverSSLProfiles, nil
-}
-
-// GetServerSSLProfile gets a server-ssl profile by name. Returns nil if the server-ssl profile does not exist
-func (b *BigIP) GetServerSSLProfile(name string) (*ServerSSLProfile, error) {
-	var serverSSLProfile ServerSSLProfile
-	err, ok := b.getForEntity(&serverSSLProfile, uriLtm, uriProfile, uriServerSSL, name)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return nil, nil
-	}
-
-	return &serverSSLProfile, nil
-}
-
-// CreateServerSSLProfile creates a new server-ssl profile on the BIG-IP system.
-func (b *BigIP) CreateServerSSLProfile(name string, parent string) error {
-	config := &ServerSSLProfile{
-		Name:         name,
-		DefaultsFrom: parent,
-	}
-
-	return b.post(config, uriLtm, uriProfile, uriServerSSL)
-}
-
-// AddServerSSLProfile adds a new server-ssl profile on the BIG-IP system.
-func (b *BigIP) AddServerSSLProfile(config *ServerSSLProfile) error {
-	return b.post(config, uriLtm, uriProfile, uriServerSSL)
-}
-
-// DeleteServerSSLProfile removes a server-ssl profile.
-func (b *BigIP) DeleteServerSSLProfile(name string) error {
-	return b.delete(uriLtm, uriProfile, uriServerSSL, name)
-}
-
-// ModifyServerSSLProfile allows you to change any attribute of a sever-ssl profile.
-// Fields that can be modified are referenced in the VirtualServer struct.
-func (b *BigIP) ModifyServerSSLProfile(name string, config *ServerSSLProfile) error {
-	return b.put(config, uriLtm, uriProfile, uriServerSSL, name)
-}
-
-// ClientSSLProfiles returns a list of client-ssl profiles.
-func (b *BigIP) ClientSSLProfiles() (*ClientSSLProfiles, error) {
-	var clientSSLProfiles ClientSSLProfiles
-	err, _ := b.getForEntity(&clientSSLProfiles, uriLtm, uriProfile, uriClientSSL)
-	if err != nil {
-		return nil, err
-	}
-
-	return &clientSSLProfiles, nil
-}
-
-// GetClientSSLProfile gets a client-ssl profile by name. Returns nil if the client-ssl profile does not exist
-func (b *BigIP) GetClientSSLProfile(name string) (*ClientSSLProfile, error) {
-	var clientSSLProfile ClientSSLProfile
-	err, ok := b.getForEntity(&clientSSLProfile, uriLtm, uriProfile, uriClientSSL, name)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return nil, nil
-	}
-
-	return &clientSSLProfile, nil
-}
-
-// CreateClientSSLProfile creates a new client-ssl profile on the BIG-IP system.
-func (b *BigIP) CreateClientSSLProfile(name string, parent string) error {
-	config := &ClientSSLProfile{
-		Name:         name,
-		DefaultsFrom: parent,
-	}
-
-	return b.post(config, uriLtm, uriProfile, uriClientSSL)
-}
-
-// AddClientSSLProfile adds a new client-ssl profile on the BIG-IP system.
-func (b *BigIP) AddClientSSLProfile(config *ClientSSLProfile) error {
-	return b.post(config, uriLtm, uriProfile, uriClientSSL)
-}
-
-// DeleteClientSSLProfile removes a client-ssl profile.
-func (b *BigIP) DeleteClientSSLProfile(name string) error {
-	return b.delete(uriLtm, uriProfile, uriClientSSL, name)
-}
-
-// ModifyClientSSLProfile allows you to change any attribute of a client-ssl profile.
-// Fields that can be modified are referenced in the ClientSSLProfile struct.
-func (b *BigIP) ModifyClientSSLProfile(name string, config *ClientSSLProfile) error {
-	return b.put(config, uriLtm, uriProfile, uriClientSSL, name)
 }
 
 // Nodes returns a list of nodes.
@@ -1118,152 +729,6 @@ func (b *BigIP) GetInternalDataGroupRecords(name string) (*[]DataGroupRecord, er
 	}
 
 	return &dataGroupRecords, nil
-}
-
-// Pools returns a list of pools.
-func (b *BigIP) Pools() (*Pools, error) {
-	var pools Pools
-	err, _ := b.getForEntity(&pools, uriLtm, uriPool)
-	if err != nil {
-		return nil, err
-	}
-
-	return &pools, nil
-}
-
-// PoolMembers returns a list of pool members for the given pool.
-func (b *BigIP) PoolMembers(name string) (*PoolMembers, error) {
-	var poolMembers PoolMembers
-	err, _ := b.getForEntity(&poolMembers, uriLtm, uriPool, name, uriPoolMember)
-	if err != nil {
-		return nil, err
-	}
-
-	return &poolMembers, nil
-}
-
-// AddPoolMember adds a node/member to the given pool. <member> must be in the form
-// of <node>:<port>, i.e.: "web-server1:443".
-func (b *BigIP) AddPoolMember(pool, member string) error {
-	config := &poolMember{
-		Name: member,
-	}
-
-	return b.post(config, uriLtm, uriPool, pool, uriPoolMember)
-}
-
-// GetPoolMember returns the details of a member in the specified pool.
-func (b *BigIP) GetPoolMember(pool string, member string) (*PoolMember, error) {
-	var poolMember PoolMember
-	err, ok := b.getForEntity(&poolMember, uriLtm, uriPool, pool, uriPoolMember, member)
-
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return nil, nil
-	}
-
-	return &poolMember, nil
-}
-
-// CreatePoolMember creates a pool member for the specified pool.
-func (b *BigIP) CreatePoolMember(pool string, config *PoolMember) error {
-	return b.post(config, uriLtm, uriPool, pool, uriPoolMember)
-}
-
-// ModifyPoolMember will update the configuration of a particular pool member.
-func (b *BigIP) ModifyPoolMember(pool string, config *PoolMember) error {
-	member := config.FullPath
-	// These fields are not used when modifying a pool member; so omit them.
-	config.Name = ""
-	config.Partition = ""
-	config.FullPath = ""
-
-	// This cannot be modified for an existing pool member.
-	config.Address = ""
-
-	return b.put(config, uriLtm, uriPool, pool, uriPoolMember, member)
-}
-
-// UpdatePoolMembers does a replace-all-with for the members of a pool.
-func (b *BigIP) UpdatePoolMembers(pool string, pm *[]PoolMember) error {
-	config := &poolMembers{
-		Members: *pm,
-	}
-	return b.put(config, uriLtm, uriPool, pool)
-}
-
-// RemovePoolMember removes a pool member from the specified pool.
-func (b *BigIP) RemovePoolMember(pool string, config *PoolMember) error {
-	member := config.FullPath
-	return b.delete(uriLtm, uriPool, pool, uriPoolMember, member)
-}
-
-// DeletePoolMember removes a member from the given pool. <member> must be in the form
-// of <node>:<port>, i.e.: "web-server1:443".
-func (b *BigIP) DeletePoolMember(pool string, member string) error {
-	return b.delete(uriLtm, uriPool, pool, uriPoolMember, member)
-}
-
-// PoolMemberStatus changes the status of a pool member. <state> can be either
-// "enable" or "disable". <member> must be in the form of <node>:<port>,
-// i.e.: "web-server1:443".
-func (b *BigIP) PoolMemberStatus(pool string, member string, state string) error {
-	config := &Node{}
-
-	switch state {
-	case "enable":
-		// config.State = "unchecked"
-		config.Session = "user-enabled"
-	case "disable":
-		// config.State = "unchecked"
-		config.Session = "user-disabled"
-		// case "offline":
-		// 	config.State = "user-down"
-		// 	config.Session = "user-disabled"
-	}
-
-	return b.put(config, uriLtm, uriPool, pool, uriPoolMember, member)
-}
-
-// CreatePool adds a new pool to the BIG-IP system by name.
-func (b *BigIP) CreatePool(name string) error {
-	config := &Pool{
-		Name: name,
-	}
-
-	return b.post(config, uriLtm, uriPool)
-}
-
-// AddPool creates a new pool on the BIG-IP system.
-func (b *BigIP) AddPool(config *Pool) error {
-	return b.post(config, uriLtm, uriPool)
-}
-
-// Get a Pool by name. Returns nil if the Pool does not exist
-func (b *BigIP) GetPool(name string) (*Pool, error) {
-	var pool Pool
-	err, ok := b.getForEntity(&pool, uriLtm, uriPool, name)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return nil, nil
-	}
-
-	return &pool, nil
-}
-
-// DeletePool removes a pool.
-func (b *BigIP) DeletePool(name string) error {
-	return b.delete(uriLtm, uriPool, name)
-}
-
-// ModifyPool allows you to change any attribute of a pool. Fields that
-// can be modified are referenced in the Pool struct.
-func (b *BigIP) ModifyPool(name string, config *Pool) error {
-	return b.put(config, uriLtm, uriPool, name)
 }
 
 // VirtualServers returns a list of virtual servers.
@@ -1481,50 +946,6 @@ func (b *BigIP) AddMonitorToPool(monitor, pool string) error {
 	}
 
 	return b.put(config, uriLtm, uriPool, pool)
-}
-
-// IRules returns a list of irules
-func (b *BigIP) IRules() (*IRules, error) {
-	var rules IRules
-	err, _ := b.getForEntity(&rules, uriLtm, uriIRule)
-	if err != nil {
-		return nil, err
-	}
-
-	return &rules, nil
-}
-
-// IRule returns information about the given iRule.
-func (b *BigIP) IRule(name string) (*IRule, error) {
-	var rule IRule
-	err, ok := b.getForEntity(&rule, uriLtm, uriIRule, name)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return nil, nil
-	}
-	return &rule, nil
-}
-
-// CreateIRule creates a new iRule on the system.
-func (b *BigIP) CreateIRule(name, rule string) error {
-	irule := &IRule{
-		Name: name,
-		Rule: rule,
-	}
-	return b.post(irule, uriLtm, uriIRule)
-}
-
-// DeleteIRule removes an iRule from the system.
-func (b *BigIP) DeleteIRule(name string) error {
-	return b.delete(uriLtm, uriIRule, name)
-}
-
-// ModifyIRule updates the given iRule with any changed values.
-func (b *BigIP) ModifyIRule(name string, irule *IRule) error {
-	irule.Name = name
-	return b.put(irule, uriLtm, uriIRule, name)
 }
 
 func (b *BigIP) Policies() (*Policies, error) {
